# WebApi/WebApi/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/amazon/upload-pdf": {
      "post": {
        "tags": [
          "Amazon"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/amazon/list-pdfs": {
      "get": {
        "tags": [
          "Amazon"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/amazon/buscar": {
      "get": {
        "tags": [
          "Amazon"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "nombre",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/amazon/descargar": {
      "get": {
        "tags": [
          "Amazon"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "nombre",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/cliente/listar": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/cliente/listarxid": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "codigo",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/cliente/guardar": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Cliente": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "edad": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        }
      }
    }
  }
}